bugs fix

+++ app/Http/Controllers/ProfileController.php
@@ -22,13 +22,14 @@ class ProfileController extends Controller
     {
         $data = $request->all();
         $str = Str::random(100);
+        $id = auth()->user()->id;
 
         $validate = validator::make($data, [
             'nama' => 'required|max:60',
             'email' => 'required|email:rfc,dns',
             'no_telp' => 'required|regex:/^08[0-9]{9,11}$/',
-            'foto' => 'required|mimes:jpg,jpeg,png',
-            'password' => 'required|min:8',
+            // 'foto' => 'required|mimes:jpg,jpeg,png',
+            // 'password' => 'min:8',
             'alamat' => 'required|max:255',
         ]);
 
@@ -36,13 +37,21 @@ class ProfileController extends Controller
             return back()->withErrors($validate);
         }
 
-        $user = User::find(auth()->user()->id);
+        $user = User::find($id);
 
-        if ($user->foto) {
-            Storage::disk('public')->delete($user->foto);
+        if (!$user) {
+            return back()->with('error', 'Data user tidak ditemukan.');
         }
 
-        $imagePath = $request->file('foto')->store('profile', 'public');
+        if ($request->file('foto') == null) {
+            unset($data['foto']);
+        } else {
+            if ($user->foto) {
+                Storage::disk('public')->delete($user->foto);
+            }
+            $imagePath = $request->file('foto')->store('profile', 'public');
+            $data['foto'] = $imagePath;
+        }
 
         if ($data['email'] != $user->email) {
             $validate = validator::make($data, [
@@ -61,7 +70,7 @@ class ProfileController extends Controller
             ];
             $data['active'] = null;
             Mail::to($request->email)->send(new MailSend($details));
-            $user = User::find(auth()->user()->id)->update([
+            $user = User::find($id)->update([
                 'nama' => $data['nama'],
                 'email' => $data['email'],
                 'no_telp' => $data['no_telp'],
@@ -72,17 +81,12 @@ class ProfileController extends Controller
                 'active' => $data['active']
             ]);
         } else {
-            $user = User::find(auth()->user()->id)->update([
-                'nama' => $data['nama'],
-                'email' => $data['email'],
-                'no_telp' => $data['no_telp'],
-                'alamat' => $data['alamat'],
-                'foto' => $imagePath,
-                'password' => $data['password']
-            ]);
+            if ($data['password'] == null) {
+                unset($data['password']);
+            }
+
+            $user = User::find($id)->update($data);
         }
-
-
 
         return redirect('user')->with('success', 'Update data diri berhasil.');
     }
